requirements.txt
streamlit
groq
edge-tts
python-dotenv
SpeechRecognition
pyaudio
pydub
sounddevice
soundfile
numpy
plotly
